+++ world/geowave.hello/INTERNAL_ADAPTER/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-000008
+MANIFEST-000010
