+++ world/geowave.hello/INTERNAL_ADAPTER/OPTIONS-000013
@@ -1,0 +1,147 @@
+# This is a RocksDB option file.
+#
+# For detailed file format spec, please refer to the example file
+# in examples/rocksdb_option_file_example.ini
+#
+
+[Version]
+  rocksdb_version=5.15.10
+  options_file_version=1.1
+
+[DBOptions]
+  skip_log_error_on_recovery=false
+  allow_fallocate=true
+  avoid_flush_during_shutdown=false
+  dump_malloc_stats=false
+  advise_random_on_open=true
+  WAL_ttl_seconds=0
+  use_direct_reads=false
+  allow_2pc=false
+  allow_mmap_reads=false
+  random_access_max_buffer_size=1048576
+  allow_mmap_writes=false
+  use_direct_io_for_flush_and_compaction=false
+  allow_concurrent_memtable_write=true
+  delayed_write_rate=16777216
+  error_if_exists=false
+  create_if_missing=true
+  compaction_readahead_size=0
+  create_missing_column_families=false
+  new_table_reader_for_compaction_inputs=false
+  enable_thread_tracking=false
+  is_fd_close_on_exec=true
+  paranoid_checks=true
+  fail_if_options_file_error=false
+  skip_stats_update_on_db_open=false
+  max_manifest_file_size=1073741824
+  enable_pipelined_write=false
+  use_adaptive_mutex=false
+  preserve_deletes=false
+  log_file_time_to_roll=0
+  use_fsync=false
+  wal_recovery_mode=kPointInTimeRecovery
+  max_background_jobs=2
+  max_background_compactions=-1
+  writable_file_max_buffer_size=1048576
+  delete_obsolete_files_period_micros=21600000000
+  base_background_compactions=-1
+  max_background_flushes=-1
+  max_log_file_size=0
+  max_file_opening_threads=1
+  max_open_files=5000
+  table_cache_numshardbits=6
+  bytes_per_sync=0
+  db_write_buffer_size=0
+  keep_log_file_num=1000
+  recycle_log_file_num=0
+  manifest_preallocation_size=4194304
+  write_thread_slow_yield_usec=3
+  db_log_dir=
+  wal_dir=./world\\geowave.hello/INTERNAL_ADAPTER
+  max_subcompactions=1
+  WAL_size_limit_MB=0
+  max_total_wal_size=0
+  write_thread_max_yield_usec=100
+  wal_bytes_per_sync=0
+  stats_dump_period_sec=600
+  enable_write_thread_adaptive_yield=true
+  access_hint_on_compaction_start=NORMAL
+  info_log_level=INFO_LEVEL
+  avoid_flush_during_recovery=false
+  allow_ingest_behind=false
+  two_write_queues=false
+  manual_wal_flush=false
+  
+
+[CFOptions "default"]
+  level0_file_num_compaction_trigger=4
+  merge_operator=nullptr
+  report_bg_io_stats=false
+  max_bytes_for_level_base=10485760
+  disable_auto_compactions=false
+  max_bytes_for_level_multiplier=10.000000
+  inplace_update_support=false
+  num_levels=7
+  target_file_size_base=2097152
+  paranoid_file_checks=false
+  level_compaction_dynamic_level_bytes=false
+  compression_per_level=
+  optimize_filters_for_hits=false
+  min_write_buffer_number_to_merge=1
+  memtable_prefix_bloom_size_ratio=0.000000
+  force_consistency_checks=false
+  soft_pending_compaction_bytes_limit=268435456
+  hard_pending_compaction_bytes_limit=1073741824
+  max_compaction_bytes=52428800
+  max_sequential_skip_in_iterations=8
+  level0_slowdown_writes_trigger=20
+  level0_stop_writes_trigger=36
+  compaction_filter_factory=nullptr
+  max_write_buffer_number=2
+  bloom_locality=0
+  max_write_buffer_number_to_maintain=0
+  target_file_size_multiplier=1
+  memtable_insert_with_hint_prefix_extractor=nullptr
+  arena_block_size=262144
+  prefix_extractor=nullptr
+  table_factory=BlockBasedTable
+  inplace_update_num_locks=10000
+  max_successive_merges=0
+  memtable_huge_page_size=0
+  compaction_style=kCompactionStyleLevel
+  write_buffer_size=2097152
+  max_bytes_for_level_multiplier_additional=1:1:1:1:1:1:1
+  compression=kNoCompression
+  bottommost_compression=kDisableCompressionOption
+  comparator=leveldb.BytewiseComparator
+  memtable_factory=SkipListFactory
+  compaction_options_universal={allow_trivial_move=false;max_size_amplification_percent=200;size_ratio=1;stop_style=kCompactionStopStyleTotalSize;min_merge_width=2;compression_size_percent=-1;max_merge_width=4294967295;}
+  compaction_filter=nullptr
+  compaction_pri=kByCompensatedSize
+  compaction_options_fifo={allow_compaction=false;max_table_files_size=1073741824;ttl=0;}
+  ttl=0
+  
+[TableOptions/BlockBasedTable "default"]
+  pin_top_level_index_and_filter=true
+  flush_block_policy_factory=FlushBlockBySizePolicyFactory
+  block_size=4096
+  cache_index_and_filter_blocks=false
+  cache_index_and_filter_blocks_with_high_priority=false
+  hash_index_allow_collision=true
+  pin_l0_filter_and_index_blocks_in_cache=false
+  index_type=kBinarySearch
+  checksum=kCRC32c
+  no_block_cache=false
+  block_size_deviation=10
+  block_restart_interval=16
+  index_block_restart_interval=1
+  metadata_block_size=4096
+  partition_filters=false
+  filter_policy=nullptr
+  whole_key_filtering=true
+  format_version=2
+  verify_compression=false
+  read_amp_bytes_per_bit=8589934592
+  block_align=false
+  enable_index_compression=true
+
